services/scraper-service/requirements.txt
# gRPC dependencies
grpcio==1.59.0
grpcio-tools==1.59.0

# Schema validation
marshmallow==3.20.1

# Environment variables
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# Web scraping
beautifulsoup4==4.12.2

# Google services
google==3.0.0
google-genai==0.3.0

# Database
pymongo==4.5.0